pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tmrt"
version = "0.1.0"
authors = [
    {name = "Muzan Sano", email = "northerntribe@example.com"},
]
description = "Token-Manifold Red Teaming for LLM Safety Research"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "torch>=2.2.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "deap>=1.3.3",
    "scikit-learn>=1.4.0",
    "unidecode>=1.3.7",
    "ftfy>=6.2.0",
    "pyyaml>=6.0.1",
    "jupyter>=1.0.0",
    "openai>=1.35.0",
    "datasets>=2.19.0",
    "sentence-transformers>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/northerntribe/tmrt"
"Bug Tracker" = "https://github.com/northerntribe/tmrt/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
